fixed rating int to float 2.0

+++ app/main.py
@@ -15,8 +15,8 @@ class CarWashStation:
     def __init__(self,
                  distance_from_city_center: int,
                  clean_power: int,
-                 average_rating: int,
-                 count_of_ratings: float) -> None:
+                 average_rating: float,
+                 count_of_ratings: int) -> None:
         self.distance_from_city_center = distance_from_city_center
         self.clean_power = clean_power
         self.average_rating = average_rating
